requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
future==0.18.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
simplejson==3.18.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.2.0
